AltairZ80: Moves console port from MMD4 to MMDC and adds MMDP device
This PR moves the console device from MMD4 to its own MMDC device. It also
adds the MMDP parallel port device. The MMDM device can now be disabled
if access to the SIMH pseudo device is needed.

This PR also adds the ability to set a drive as single or double-sided.